Extract Class - TextReport
Extracted the TextReport class to contain the report logic that was
gumming up the statement method. Still has a callback to TextReport that
would be nice to get rid of, but can't think of a way to. Rubocop is
satisfied with the exception of ABC violation on price per rental still.