extract price and frequent_renter_points
The #price and #frequent_renter_points in Movie can be moved to a module
that allows those behaviors to be added in dynamically. Ideally the
pricing strategy would be passed in during initialize, but since we are
refactoring that doesn't allow us to change the method signature. In a
subsequent commit I will dry up the modules.