add custom RSpec matcher
This custom matcher allows comparison of the text report formatter. It
wants a ReportData object as an argument to match_text_report. This
convenince object (defined currently in spec_helper) allows you to set
the expected variables that are in the report, as well as a method to
add the different Movies that are part of the Rental.